issue/168: decompose ChannelDispatcher into partial-class facets
Splits the 1047-LOC ChannelDispatcher into 8 partial-class files (each
< 250 LOC) by responsibility: Lifecycle, Loop, Inbox, Sinks, Operator,
Packet, WorkItem. No API changes, no test changes.

Closes #168.
